[minor] rename files and cosmetic changes
