+++ tests/qlan/test_qlan_measurement.py
@@ -1,9 +1,9 @@
 
-from sequence.qlan.qlan_orchestrator import QlanOrchestratorNode
-from sequence.qlan.qlan_client import QlanClientNode
+from sequence.qlan.orchestrator import QlanOrchestratorNode
+from sequence.qlan.client import QlanClientNode
 from sequence.kernel.timeline import Timeline
 from sequence.components.optical_channel import ClassicalChannel
-from sequence.qlan.qlan_measurement import QlanMeasurementMsgType, QlanB0MsgType
+from sequence.qlan.measurement import QlanMeasurementMsgType, QlanB0MsgType
 
 def pair_protocol(orchestrator: QlanOrchestratorNode, clients: list[QlanClientNode]):
     
@@ -22,20 +22,13 @@ def pair_protocol(orchestrator: QlanOrchestratorNode, clients: list[QlanClientNo
 def test_send_outcome_messages_z_0():
     tl = Timeline()
 
-    test_client1 = QlanClientNode(name='test_client1',
-                            tl=tl, 
-                            num_local_memories=1)
-    test_client2 = QlanClientNode(name='test_client2',
-                            tl=tl, 
-                            num_local_memories=1)
+    test_client1 = QlanClientNode(name='test_client1', tl=tl, num_local_memories=1)
+    test_client2 = QlanClientNode(name='test_client2', tl=tl, num_local_memories=1)
     
     test_memo1 = test_client1.get_components_by_type("Memory")[0]
     test_memo2 = test_client2.get_components_by_type("Memory")[0]
 
-    test_orch = QlanOrchestratorNode("test_orch",
-                                tl, 
-                                num_local_memories=1, 
-                                remote_memories=[test_memo1, test_memo2])
+    test_orch = QlanOrchestratorNode("test_orch", tl, num_local_memories=1, remote_memories=[test_memo1, test_memo2])
     test_orch.set_seed(2)
     test_orch.update_bases('z')
     
@@ -58,20 +51,13 @@ def test_send_outcome_messages_z_0():
 def test_send_outcome_messages_z_1():
     tl = Timeline()
 
-    test_client1 = QlanClientNode(name='test_client1',
-                            tl=tl, 
-                            num_local_memories=1)
-    test_client2 = QlanClientNode(name='test_client2',
-                            tl=tl, 
-                            num_local_memories=1)
+    test_client1 = QlanClientNode(name='test_client1', tl=tl, num_local_memories=1)
+    test_client2 = QlanClientNode(name='test_client2', tl=tl, num_local_memories=1)
     
     test_memo1 = test_client1.get_components_by_type("Memory")[0]
     test_memo2 = test_client2.get_components_by_type("Memory")[0]
 
-    test_orch = QlanOrchestratorNode("test_orch",
-                                tl, 
-                                num_local_memories=1, 
-                                remote_memories=[test_memo1, test_memo2])
+    test_orch = QlanOrchestratorNode("test_orch", tl, num_local_memories=1, remote_memories=[test_memo1, test_memo2])
     test_orch.set_seed(0)
     test_orch.update_bases('z')
     
@@ -95,20 +81,13 @@ def test_send_outcome_messages_z_1():
 def test_send_outcome_messages_y_0():
     tl = Timeline()
 
-    test_client1 = QlanClientNode(name='test_client1',
-                            tl=tl, 
-                            num_local_memories=1)
-    test_client2 = QlanClientNode(name='test_client2',
-                            tl=tl, 
-                            num_local_memories=1)
+    test_client1 = QlanClientNode(name='test_client1', tl=tl, num_local_memories=1)
+    test_client2 = QlanClientNode(name='test_client2', tl=tl, num_local_memories=1)
     
     test_memo1 = test_client1.get_components_by_type("Memory")[0]
     test_memo2 = test_client2.get_components_by_type("Memory")[0]
 
-    test_orch = QlanOrchestratorNode("test_orch",
-                                tl, 
-                                num_local_memories=1, 
-                                remote_memories=[test_memo1, test_memo2])
+    test_orch = QlanOrchestratorNode("test_orch", tl, num_local_memories=1, remote_memories=[test_memo1, test_memo2])
     test_orch.set_seed(2)
     test_orch.update_bases('y')
     
@@ -131,20 +110,13 @@ def test_send_outcome_messages_y_0():
 def test_send_outcome_messages_y_1():
     tl = Timeline()
 
-    test_client1 = QlanClientNode(name='test_client1',
-                            tl=tl, 
-                            num_local_memories=1)
-    test_client2 = QlanClientNode(name='test_client2',
-                            tl=tl, 
-                            num_local_memories=1)
+    test_client1 = QlanClientNode(name='test_client1', tl=tl, num_local_memories=1)
+    test_client2 = QlanClientNode(name='test_client2', tl=tl, num_local_memories=1)
     
     test_memo1 = test_client1.get_components_by_type("Memory")[0]
     test_memo2 = test_client2.get_components_by_type("Memory")[0]
 
-    test_orch = QlanOrchestratorNode("test_orch",
-                                tl, 
-                                num_local_memories=1, 
-                                remote_memories=[test_memo1, test_memo2])
+    test_orch = QlanOrchestratorNode("test_orch", tl, num_local_memories=1, remote_memories=[test_memo1, test_memo2])
     test_orch.set_seed(0)
     test_orch.update_bases('y')
     
@@ -167,20 +139,13 @@ def test_send_outcome_messages_y_1():
 def test_send_outcome_messages_x_0():
     tl = Timeline()
 
-    test_client1 = QlanClientNode(name='test_client1',
-                            tl=tl, 
-                            num_local_memories=1)
-    test_client2 = QlanClientNode(name='test_client2',
-                            tl=tl, 
-                            num_local_memories=1)
+    test_client1 = QlanClientNode(name='test_client1', tl=tl, num_local_memories=1)
+    test_client2 = QlanClientNode(name='test_client2', tl=tl, num_local_memories=1)
     
     test_memo1 = test_client1.get_components_by_type("Memory")[0]
     test_memo2 = test_client2.get_components_by_type("Memory")[0]
 
-    test_orch = QlanOrchestratorNode("test_orch",
-                                tl, 
-                                num_local_memories=1, 
-                                remote_memories=[test_memo1, test_memo2])
+    test_orch = QlanOrchestratorNode("test_orch", tl, num_local_memories=1, remote_memories=[test_memo1, test_memo2])
     test_orch.set_seed(2)
     test_orch.update_bases('x')
     
@@ -203,20 +168,13 @@ def test_send_outcome_messages_x_0():
 def test_send_outcome_messages_x_1():
     tl = Timeline()
 
-    test_client1 = QlanClientNode(name='test_client1',
-                            tl=tl, 
-                            num_local_memories=1)
-    test_client2 = QlanClientNode(name='test_client2',
-                            tl=tl, 
-                            num_local_memories=1)
+    test_client1 = QlanClientNode(name='test_client1', tl=tl, num_local_memories=1)
+    test_client2 = QlanClientNode(name='test_client2', tl=tl, num_local_memories=1)
     
     test_memo1 = test_client1.get_components_by_type("Memory")[0]
     test_memo2 = test_client2.get_components_by_type("Memory")[0]
 
-    test_orch = QlanOrchestratorNode("test_orch",
-                                tl, 
-                                num_local_memories=1, 
-                                remote_memories=[test_memo1, test_memo2])
+    test_orch = QlanOrchestratorNode("test_orch", tl, num_local_memories=1, remote_memories=[test_memo1, test_memo2])
     test_orch.set_seed(0)
     test_orch.update_bases('x')
